% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_ppmf12_path}
\alias{add_ppmf12_path}
\title{Add ppmf12 path to Renviron}
\usage{
add_ppmf12_path(path, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{path}{path where ppmf12 data is stored}

\item{overwrite}{Defaults to FALSE. Should existing ppmf12 in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Should ppmf12 be added to '~/.Renviron' file?}
}
\value{
path, invisibly
}
\description{
Add ppmf12 path to Renviron
}
\examples{
\dontrun{
tp <- tempfile(fileext = '.csv')
add_ppmf12_path(tp)
path12 <- Sys.getenv('path12')
}

}
\concept{save}
