% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_ppmf19_path}
\alias{add_ppmf19_path}
\title{Add ppmf19 path to Renviron}
\usage{
add_ppmf19_path(path, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{path}{path where ppmf19 data is stored}

\item{overwrite}{Defaults to FALSE. Should existing ppmf19 in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Should ppmf19 be added to '~/.Renviron' file?}
}
\value{
path, invisibly
}
\description{
Add ppmf19 path to Renviron
}
\examples{
\dontrun{
tp <- tempfile(fileext = '.csv')
add_ppmf19_path(tp)
path19 <- Sys.getenv('path19')
}

}
\concept{save}
