% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_ppmf19r_path}
\alias{add_ppmf19r_path}
\title{Add ppmf19r path to Renviron}
\usage{
add_ppmf19r_path(path, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{path}{path where ppmf19r data is stored}

\item{overwrite}{Defaults to FALSE. Should existing ppmf19 in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Should ppmf19r be added to '~/.Renviron' file?}
}
\value{
path, invisibly
}
\description{
Path for the 19.61 \emph{r}eplication in 2023.
}
\examples{
\dontrun{
tp <- tempfile(fileext = '.csv')
add_ppmf19r_path(tp)
path19 <- Sys.getenv('path19')
}

}
\concept{save}
