% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_path.R
\name{add_ppmf4_path}
\alias{add_ppmf4_path}
\title{Add ppmf4 path to Renviron}
\usage{
add_ppmf4_path(path, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{path}{path where ppmf4 data is stored}

\item{overwrite}{Defaults to FALSE. Should existing ppmf4 in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Should ppmf4 be added to '~/.Renviron' file?}
}
\value{
path, invisibly
}
\description{
Add ppmf4 path to Renviron
}
\examples{
\dontrun{
tp <- tempfile(fileext = '.csv')
add_ppmf4_path(tp)
path4 <- Sys.getenv('path4')
}

}
\concept{save}
