% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoid.R
\name{breakdown_geoid}
\alias{breakdown_geoid}
\title{Breakdown GEOID into Components}
\usage{
breakdown_geoid(ppmf, GEOID = GEOID)
}
\arguments{
\item{ppmf}{tibble of ppmf data}

\item{GEOID}{Column in ppmf with GEOID. Default is \code{GEOID}.}
}
\value{
tibble. ppmf with columns added for state, county, tract, block group, and/or block
}
\description{
Breakdown GEOID into Components
}
\examples{
data(ppmf_ex)
ppmf_ex <- ppmf_ex \%>\% add_geoid()
ppmf_ex <- ppmf_ex \%>\% breakdown_geoid()
}
\concept{geoid}
