% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{overwrite_hisp_race}
\alias{overwrite_hisp_race}
\title{Overwrite Races with Hispanic}
\usage{
overwrite_hisp_race(ppmf, race = CENRACE, hisp = CENHISP)
}
\arguments{
\item{ppmf}{tibble of ppmf data}

\item{race}{Column in ppmf containing race codes}

\item{hisp}{Column in ppmf containing 1 for Not Hispanic and 2 for Hispanic}
}
\value{
tibble with race column entries replaced if the individual is Hispanic
}
\description{
Overwrite Races with Hispanic
}
\examples{
data(ppmf_ex)
ppmf_ex |> replace_race() |> overwrite_hisp_race()
}
\concept{basic}
