% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ppmf_ex}
\alias{ppmf_ex}
\title{Example PPMF Data}
\usage{
data('ppmf_ex')
}
\value{
tibble with sample ppmf data
}
\description{
Includes Perry County, Alabama PPMF data from the April 28, 2021 PPMF data release.
This is a subset taken from the 12-2 P data.

As each observation is a person, this does not cover every block in the county and
due to DAS, not every block with population appears in this data.
}
\examples{
data('ppmf_ex')
}
