% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{races}
\alias{races}
\title{Race Classifications}
\usage{
data('races')
}
\value{
tibble with three columns
\itemize{
\item code: the two digit code used to code races
\item desc: the description of the races
\item group: the summary group used
}
}
\description{
This data includes the basic race classifications used for redistricting to
get to an easier to work with set of values. This does not include \code{hisp} grouping
which is controlled separately by race within the census
}
\examples{
data('races')

}
\concept{data}
