% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.R
\name{replace_race}
\alias{replace_race}
\title{Replace Race Categories}
\usage{
replace_race(ppmf, race = CENRACE)
}
\arguments{
\item{ppmf}{tibble of ppmf data}

\item{race}{Column in ppmf containing race codes}
}
\value{
tibble with race column replaced by simpler racial classifications
}
\description{
Replaces the Census's numeric categories for race with less specific racial
classifications, typically useful for redistricting purposes.
}
\examples{
data(ppmf_ex)
ppmf_ex \%>\% replace_race()
}
