% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{states}
\alias{states}
\title{State Rows}
\usage{
data('states')
}
\value{
tibble with sample ppmf data
}
\description{
This data includes the 52 geographies (50 states plus D.C. and P.R.). Within the
2010 PPMF, skip and n_max indicate the relevant rows for a geography.
}
\examples{
data('states')
}
